updated after refactoring meeting 16th March
