+++ hangman_code/functions_for_play_game/guessed_letters.py
@@ -1,3 +1,4 @@
+
 def used_letters_function(letter_list, new_text):
     letter_list [:]= [item.lower() for item in letter_list]
             
@@ -7,11 +8,7 @@ def used_letters_function(letter_list, new_text):
     return letter_list
     print(letter_list)
 
-def guessed_words_function(guessed_word, words_guessed):
-#This will make a list of previously guessed words
-# Returns a list of guessed words
-# Should be to lower, trim white space etc
-        return None
+
