+++ hangman_code/functions_for_play_game/make_guess.py
@@ -24,3 +24,13 @@ def Make_guess(letter, word, guessed_word):
                 "message": "Bad guess you lemon",
                 "word_progress": word_with_guessed_letters
                 }
+
+## ADD THESE IN AS UPDATES IN MAKE GUESS
+
+def remaining_attempts_function(attempts_remaining):
+       #This function will update the number of remaining attempts left
+       return attempts_remaining
+
+def update_score_function(current_score):
+       #This function will update the score
+       return current_score
