--- hangman_code/functions_for_play_game/other_game_data.py
@@ -1,3 +1,0 @@
-def index ():
-    game_index = ""
-    return game_index
