--- hangman_code/functions_for_play_game/scores_and_attempts_function.py
@@ -1,21 +1,0 @@
-def remaining_attempts_function(attempts_remaining):
-       #This function will update the number of remaining attempts left
-       return attempts_remaining
-
-def update_score_function(current_score):
-       #This function will update the score
-       return current_score
-
-def update_cunulative_score_function(cumulative_score):
-       #This function will update the cumulative score
-       return cumulative_score
-
-def update_cunulative_number_of_games_played(number_of_games_played):
-       #This function will update the cumulative number of games played
-       return number_of_games_played
-
-def update_cunulative_number_of_games_won(number_of_games_won):
-       #This function will update the cumulative number of games played
-       return number_of_games_won
-
-
