--- hangman_code/static.py
@@ -1,13 +1,0 @@
-"Contains Static Assets such as the CSS, JavaScript, "
-"and images used by templates"
-"Relationship to templates.py: recieves a request "
-"from templates and provides the asset"
-
-class Static():
-
-    def _init_(self):
-        #Contains the variables
-        return None
-    
-    def static_assets():
-        return None
